modified bookmarks to work with not logged in users
Using the notion of guest users, a new user object is created for
any visitor as soon as they visit the web site. Therefore, bookmarks and saved searches can be used immediately.

If the user then logs in to an existing or new account, any bookmarks or saved searches are then merged.

References #447

still to do:
* testing, what's that?
* do we remove folders completely?
* testing what happens when there's not an authentication framework